Add files via upload
This game was created using pygame. The game source was created by referring to the free lectures on YouTube. Korea created a system called 'GAP' for the safety management of agricultural products, and I made a game for promotion while working as a supporter of this system.

Good Agricultural Practices (GAP) is a system that manages and certifies factors that threaten food safety, such as pesticides, from production farms to consumers. The game was organized by attacking and eliminating elements that threaten GAP corn with missiles. Use the rudder and space bar to control, and if you miss a total of three times, the game will end.